Check for LinuxRT
Changed the check IF 903x to check for LinuxRT, then checks whether it's
a 903x or 906x.

Inserted case structure for using the Set System Image (file) call to
LinuxRT instead of the folder version like normal. Both still use the
folder version as I had trouble getting it to work for now.