Updating Global Timeout
There are three places we see timeout values.  Two are defaulted to 3
and 4 seconds respectively.  The third (NI-Auth credentials) has a 30
second default timeout.  In some corner cases, increasing the 3 and 4
second timeout resolves the issue.

Modified code to add a global timeout rather than wiring as these were
the only values used anywhere.  If they need to be updated later in
source, they can use the global timeout to make the updates.

For the 3s timeouts, those were updated to 4s.  For the 30s timeout, the
global is updated temporarily to send the desired timeout to the open
session.  After this subVI is complete (rad_Modify and Deploy
RT-INI.vi), the global timeout is set back to the user's desired value.